Major changes
1. It allows `wdi` to access quarterly data from the WorldBankAPI, and not
just yearly data.

2. It addes two keywords: "all" and "all_countries" to the wdi function
that returns the series for all countries/regions and for all countries
respectively.

3. It solves a problem when calling the worldbank api for a data range
that is a strict superset of the available data. For example:

https://api.worldbank.org/v2/countries/US/indicators/NY.GDP.PCAP.CD?format=json&per_page=25000&date=1950:2021

returns empty.

This is done by always collecting all of the available data for the
indicator, and then filtering it accordingly to the requested dates.

4. It introduces a new dependency, Dates, which facilitates dealing with
quarterly data.

5. The returned DataFrame from wdi now includes a frequency column,
and a date column (in addition to the year).